Refactor more Network connect functions to subclasses
